Allow to compile with griffin or common compilation in PS2
