+++ Makefile.griffin
@@ -1197,6 +1197,10 @@ ifeq ($(HAVE_FILTERS_BUILTIN), 1)
    CFLAGS += -DHAVE_FILTERS_BUILTIN
 endif
 
+ifeq ($(HAVE_WINDOW_OFFSET), 1)
+   CFLAGS += -DHAVE_WINDOW_OFFSET
+endif
+
 ifeq ($(HAVE_THREADS), 1)
    CFLAGS += -DHAVE_THREADS
